// client/src/app/@core/enums/loyalties/loyalties.enum.ts
export enum ActionHandleLoyalty {
    INCREASE = "increase",
    DECREASE = "decrease",
    DEFAULT = "default",
    DOUBLE = 'double',
    BONUS = 'bonus'
}

export enum LoyaltyRank {
    BRONZE = 0,
    SILVER = 1000,
    GOLD = 5000,
    DIAMOND = 10000,
}

export enum LoyaltyRankLabel {
    BRONZE = 'Đồng',
    SILVER = 'Bạc',
    GOLD = 'Vàng',
    DIAMOND = 'Kim cương',
}